Fix packaging/making with npm
The code was running 'npm prune', but that doesn't do anything if no dependencies are already installed. What we really want is 'npm install --production' and I also added '--no-bin-links' to match the yarn behavior. Also, added end-to-end unit tests running with npm in addition to the ones running with yarn.

Fixes #308, #309